Revert "Отдельное окно предпросмотра"
