+++ nfprogress/ProgressShareImage.swift
@@ -10,15 +10,12 @@ public typealias OSImage = NSImage
 #endif
 
 /// Size of the exported progress image in points.
-let shareImageSize: CGFloat = 500
-
+private let shareImageSize: CGFloat = 500
 
 /// Snapshot of ``ProgressCircleView`` without animations.
 private struct ProgressCircleSnapshotView: View {
     var project: WritingProject
-    var size: CGFloat
-    var ringWidth: CGFloat
-    var percentFontSize: CGFloat
+    var style: ProgressCircleStyle = .large
 
     private var progress: Double {
         guard project.goal > 0 else { return 0 }
@@ -26,7 +23,9 @@ private struct ProgressCircleSnapshotView: View {
         return min(max(value, 0), 1)
     }
 
+    private var ringWidth: CGFloat { (style == .large ? layoutStep(3) : layoutStep(2)) * 2 }
     private var color: Color { .interpolate(from: .red, to: .green, fraction: progress) }
+    private var fontToken: FontToken { style == .large ? .progressValueLarge : .progressValue }
 
     var body: some View {
         ZStack {
@@ -37,38 +36,24 @@ private struct ProgressCircleSnapshotView: View {
                 .rotationEffect(.degrees(-90))
             let percent = Int(ceil(progress * 100))
             Text("\(percent)%")
-                .font(.system(size: percentFontSize))
+                .font(.system(size: calcFontSize(token: fontToken) * 3))
                 .monospacedDigit()
                 .bold()
                 .foregroundColor(color)
         }
-        .frame(width: size, height: size)
     }
 }
 
-struct ProgressShareView: View {
+private struct ProgressShareView: View {
     var project: WritingProject
-    var circleSize: CGFloat = CGFloat(defaultShareCircleSize)
-    var ringWidth: CGFloat = CGFloat(defaultShareRingWidth)
-    var percentFontSize: CGFloat = CGFloat(defaultSharePercentSize)
-    var titleFontSize: CGFloat = CGFloat(defaultShareTitleSize)
-    var titleSpacing: CGFloat = CGFloat(defaultShareSpacing)
 
     var body: some View {
-        VStack(spacing: 0) {
-            Spacer()
-            ProgressCircleSnapshotView(project: project,
-                                       size: circleSize,
-                                       ringWidth: ringWidth,
-                                       percentFontSize: percentFontSize)
-            Spacer().frame(height: titleSpacing)
+        VStack(spacing: scaledSpacing(2)) {
+            ProgressCircleSnapshotView(project: project, style: .large)
+                .frame(width: shareImageSize * 0.7, height: shareImageSize * 0.7)
             Text(project.title)
-                .font(.system(size: titleFontSize, weight: .bold))
+                .font(.title.bold())
                 .multilineTextAlignment(.center)
-                .foregroundColor(.black)
-                .lineLimit(nil)
-                .fixedSize(horizontal: false, vertical: true)
-            Spacer()
         }
         .frame(width: shareImageSize, height: shareImageSize)
         .background(Color.white)
@@ -76,18 +61,8 @@ struct ProgressShareView: View {
 }
 
 @MainActor
-func progressShareImage(for project: WritingProject,
-                        circleSize: CGFloat = CGFloat(defaultShareCircleSize),
-                        ringWidth: CGFloat = CGFloat(defaultShareRingWidth),
-                        percentFontSize: CGFloat = CGFloat(defaultSharePercentSize),
-                        titleFontSize: CGFloat = CGFloat(defaultShareTitleSize),
-                        titleSpacing: CGFloat = CGFloat(defaultShareSpacing)) -> OSImage? {
-    let view = ProgressShareView(project: project,
-                                 circleSize: circleSize,
-                                 ringWidth: ringWidth,
-                                 percentFontSize: percentFontSize,
-                                 titleFontSize: titleFontSize,
-                                 titleSpacing: titleSpacing)
+func progressShareImage(for project: WritingProject) -> OSImage? {
+    let view = ProgressShareView(project: project)
     let renderer = ImageRenderer(content: view)
 #if swift(>=5.9)
     renderer.proposedSize = ProposedViewSize(width: shareImageSize, height: shareImageSize)
@@ -105,18 +80,8 @@ func progressShareImage(for project: WritingProject,
 
 
 @MainActor
-func progressShareURL(for project: WritingProject,
-                      circleSize: CGFloat = CGFloat(defaultShareCircleSize),
-                      ringWidth: CGFloat = CGFloat(defaultShareRingWidth),
-                      percentFontSize: CGFloat = CGFloat(defaultSharePercentSize),
-                      titleFontSize: CGFloat = CGFloat(defaultShareTitleSize),
-                      titleSpacing: CGFloat = CGFloat(defaultShareSpacing)) -> URL? {
-    guard let image = progressShareImage(for: project,
-                                         circleSize: circleSize,
-                                         ringWidth: ringWidth,
-                                         percentFontSize: percentFontSize,
-                                         titleFontSize: titleFontSize,
-                                         titleSpacing: titleSpacing) else { return nil }
+func progressShareURL(for project: WritingProject) -> URL? {
+    guard let image = progressShareImage(for: project) else { return nil }
 #if canImport(UIKit)
     guard let data = image.pngData() else { return nil }
 #else
