+++ nfprogress/ProjectDetailView.swift
@@ -30,8 +30,9 @@ struct ProjectDetailView: View {
     @State private var isEditingDeadline = false
     @FocusState private var focusedField: Field?
 #if os(iOS)
-    @State private var showingSharePreview = false
+    @State private var showingShareSheet = false
 #endif
+    @State private var shareURL: URL?
 
     /// Base spacing for history and stages sections.
     private let viewSpacing: CGFloat = scaledSpacing(2)
@@ -290,19 +291,24 @@ struct ProjectDetailView: View {
     }
 
     private func shareToolbarButton() -> some View {
-#if os(macOS)
-        Button(action: {
-            let request = SharePreviewRequest(projectID: project.id)
-            openWindow(id: "sharePreview", value: request)
-        }) {
+        Button(action: shareProgress) {
             Image(systemName: "square.and.arrow.up")
         }
         .help(settings.localized("share_progress_tooltip"))
+    }
+
+    private func shareProgress() {
+        guard let url = progressShareURL(for: project) else { return }
+        shareURL = url
+#if os(iOS)
+        showingShareSheet = true
 #else
-        Button(action: { showingSharePreview = true }) {
-            Image(systemName: "square.and.arrow.up")
+        let picker = NSSharingServicePicker(items: [url])
+        if let window = NSApplication.shared.keyWindow {
+            picker.show(relativeTo: .zero, of: window.contentView!, preferredEdge: .minY)
+        } else if let window = NSApplication.shared.windows.first {
+            picker.show(relativeTo: .zero, of: window.contentView!, preferredEdge: .minY)
         }
-        .help(settings.localized("share_progress_tooltip"))
 #endif
     }
 
@@ -519,9 +525,13 @@ struct ProjectDetailView: View {
             }
         }
 #if os(iOS)
-        .sheet(isPresented: $showingSharePreview) {
-            ProgressSharePreview(project: project)
-                .environmentObject(settings)
+        .sheet(isPresented: $showingShareSheet, onDismiss: {
+            if let url = shareURL { try? FileManager.default.removeItem(at: url) }
+            shareURL = nil
+        }) {
+            if let url = shareURL {
+                ShareSheet(items: [url])
+            }
         }
 #endif
     }
